refactor: 신고된 채팅 내용 조회 로직을 MatchingReportService로 이동
