+++ src/main/java/econo/buddybridge/report/service/MatchingReportService.java
@@ -1,12 +1,16 @@
 package econo.buddybridge.report.service;
 
+import econo.buddybridge.chat.chatmessage.dto.ChatMessageCustomPage;
+import econo.buddybridge.common.persistence.filter.annotation.WithDeletedContent;
 import econo.buddybridge.matching.entity.Matching;
+import econo.buddybridge.matching.service.MatchingRoomService;
 import econo.buddybridge.matching.service.MatchingService;
 import econo.buddybridge.member.entity.Member;
 import econo.buddybridge.member.service.MemberService;
 import econo.buddybridge.report.dto.ReportRequest;
 import econo.buddybridge.report.entity.MatchingReport;
 import econo.buddybridge.report.exception.ReportMatchingAlreadyExistsException;
+import econo.buddybridge.report.exception.ReportNotFoundException;
 import econo.buddybridge.report.repository.MatchingReportRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,6 +22,7 @@ public class MatchingReportService {
 
     private final MatchingReportRepository matchingReportRepository;
     private final MatchingService matchingService;
+    private final MatchingRoomService matchingRoomService;
     private final MemberService memberService;
 
     @Transactional
@@ -35,5 +40,19 @@ public class MatchingReportService {
                 reportRequest.reportType(),
                 reportRequest.reportReason()
         ));
+    }
+
+    @Transactional(readOnly = true)
+    @WithDeletedContent
+    public ChatMessageCustomPage getMatchingRoomMessages(Long reportId, Integer size, Long cursor) {
+        MatchingReport matchingReport = findReportByIdOrThrow(reportId);
+        Member reporter = matchingReport.getReporter();
+        Long matchingId = matchingReport.getReportedMatching().getId();
+        return matchingRoomService.getReportedMatchingRoomMessages(reporter.getId(), matchingId, size, cursor);
+    }
+
+    private MatchingReport findReportByIdOrThrow(Long reportId) {
+        return matchingReportRepository.findById(reportId)
+                .orElseThrow(() -> ReportNotFoundException.EXCEPTION);
     }
 }
